feat(table): add filtertype prop and update stories
