+++ core/components/organisms/table/__stories__/index.story.jsx
@@ -1,258 +1,321 @@
 import * as React from 'react';
-import { Table, Card } from '@/index';
+import loaderSchema from '@/components/organisms/grid/__stories__/_common_/loaderSchema';
+import data from '@/components/organisms/grid/__stories__/_common_/data';
+import { Card, Table, Button } from '@/index';
 import { AsyncTable, SyncTable } from './_common_/types';
+import { fetchData } from '@/components/organisms/grid/__stories__/_common_/fetchData';
+import { action } from '@/utils/action';
 
 export const all = () => {
-  const data = [
-    {
-      firstName: 'Brooke',
-      lastName: 'Heeran',
-      email: 'bheeran0@altervista.org',
-      gender: 'Female',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Frazer',
-      lastName: 'Cathro',
-      email: 'fcathro1@ucla.edu',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Lemmie',
-      lastName: 'Ciric',
-      email: {
-        title: 'lciric2@dmoz.org',
-        metaList: ['First', 'Second'],
-      },
-      gender: 'Male',
-      status: 'Completed',
-    },
-    {
-      firstName: 'Randy',
-      lastName: 'Boatwright',
-      email: 'rboatwright3@arstechnica.com',
-      status: 'Completed',
-      gender: 'Male',
-    },
-    {
-      firstName: 'Rolando',
-      lastName: 'Cyples',
-      email: 'rcyples4@biglobe.ne.jp',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Lem',
-      lastName: 'Males',
-      email: 'lmales5@admin.ch',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Sayres',
-      lastName: 'Adelberg',
-      email: 'sadelberg6@uol.com.br',
-      gender: 'Male',
-      status: 'Completed',
-    },
-    {
-      firstName: 'Murray',
-      lastName: 'Bravington',
-      email: 'mbravington7@drupal.org',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Jena',
-      lastName: 'Swatheridge',
-      email: 'jswatheridge8@npr.org',
-      gender: 'Female',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Annabel',
-      lastName: 'Nelsey',
-      email: 'anelsey9@google.com',
-      gender: 'Female',
-      status: 'Completed',
-    },
-  ];
-
-  const schema = [
-    {
-      name: 'firstName',
-      displayName: 'Name',
-      cellType: 'AVATAR_WITH_TEXT',
-      width: '25%',
-      translate: (a) => ({
-        title: `${a.firstName} ${a.lastName}`,
-        firstName: a.firstName,
-        lastName: a.lastName,
-      }),
-    },
-    {
-      name: 'email',
-      displayName: 'Email',
-      width: '25%',
-    },
-    {
-      name: 'gender',
-      displayName: 'Gender',
-      width: '25%',
-    },
-    {
-      name: 'status',
-      displayName: 'Status',
-      width: '25%',
-    },
-  ];
+  const selectionActionRenderer = (selectedData, selectAll) => {
+    action('selectedData', selectedData, 'selectAll', selectAll)();
+    return (
+      <div className="d-flex align-items-center">
+        <Button size="tiny" className="mr-4">
+          Delete
+        </Button>
+        <Button size="tiny">Export</Button>
+      </div>
+    );
+  };
 
   return (
-    <div className="vh-50">
+    <div className="vh-75">
       <Card className="h-100 overflow-hidden">
         <Table
-          page={1}
-          data={data}
-          schema={schema}
-          showMenu={true}
+          loaderSchema={loaderSchema}
+          fetchData={fetchData}
           withHeader={true}
-          withPagination={true}
-          pageSize={4}
+          withCheckbox={true}
+          uniqueColumnName="firstName"
+          onSelect={(rowIndex, selected, selectedList, selectAll) =>
+            action(
+              `on-select:- rowIndex: ${rowIndex} selected: ${selected} selectedList: ${JSON.stringify(
+                selectedList
+              )} selectAll: ${selectAll}`
+            )()
+          }
           headerOptions={{
             withSearch: true,
+            allowSelectAll: true,
+            selectionActionRenderer,
           }}
+          withPagination={true}
+          pageSize={5}
+          onPageChange={(newPage) => action(`on-page-change:- ${newPage}`)()}
         />
       </Card>
     </div>
   );
 };
 
-const customCode = `() => {
-  const data = [
-    {
-      firstName: 'Brooke',
-      lastName: 'Heeran',
-      email: 'bheeran0@altervista.org',
-      gender: 'Female',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Frazer',
-      lastName: 'Cathro',
-      email: 'fcathro1@ucla.edu',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Lemmie',
-      lastName: 'Ciric',
-      email: {
-        title: 'lciric2@dmoz.org',
-        metaList: ['First', 'Second'],
-      },
-      gender: 'Male',
-      status: 'Completed',
-    },
-    {
-      firstName: 'Randy',
-      lastName: 'Boatwright',
-      email: 'rboatwright3@arstechnica.com',
-      status: 'Completed',
-      gender: 'Male',
-    },
-    {
-      firstName: 'Rolando',
-      lastName: 'Cyples',
-      email: 'rcyples4@biglobe.ne.jp',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Lem',
-      lastName: 'Males',
-      email: 'lmales5@admin.ch',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Sayres',
-      lastName: 'Adelberg',
-      email: 'sadelberg6@uol.com.br',
-      gender: 'Male',
-      status: 'Completed',
-    },
-    {
-      firstName: 'Murray',
-      lastName: 'Bravington',
-      email: 'mbravington7@drupal.org',
-      gender: 'Male',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Jena',
-      lastName: 'Swatheridge',
-      email: 'jswatheridge8@npr.org',
-      gender: 'Female',
-      status: 'Failed',
-    },
-    {
-      firstName: 'Annabel',
-      lastName: 'Nelsey',
-      email: 'anelsey9@google.com',
-      gender: 'Female',
-      status: 'Completed',
-    },
-  ];
+const customCode = `
+() => {
+  const translateData = (schema, data) => {
+    let newData = data;
+
+    if (schema.translate) {
+      const translatedData = schema.translate(data);
+      newData = {
+        ...newData,
+        [schema.name]: typeof translatedData === 'object' ? {
+          ...newData[schema.name],
+          ...translatedData
+        } : translatedData
+      };
+    }
+    if (typeof newData[schema.name] !== 'object') newData[schema.name] = { title: newData[schema.name] };
+
+    return newData;
+  }
+
+  const filterData = (schema, data, filterList) => {
+    let filteredData = data;
+    if (filterList) {
+      Object.keys(filterList).forEach(schemaName => {
+        const filters = filterList[schemaName];
+        const sIndex = schema.findIndex(s => s.name === schemaName);
+        const { onFilterChange } = schema[sIndex];
+        if (filters.length && onFilterChange) {
+          filteredData = filteredData.filter(d => onFilterChange(d, filters));
+        }
+      });
+    }
+
+    return filteredData;
+  };
+
+  const sortData = (schema, data, sortingList) => {
+    const sortedData = [...data];
+    sortingList.forEach(l => {
+      const sIndex = schema.findIndex(s => s.name === l.name);
+      if (sIndex !== -1) {
+        const defaultComparator = (a, b) => {
+          const aData = translateData(schema[sIndex], a);
+          const bData = translateData(schema[sIndex], b);
+          return aData[l.name].title.localeCompare(bData[l.name].title);
+        };
+
+        const {
+          comparator = defaultComparator
+        } = schema[sIndex];
+
+        sortedData.sort(comparator);
+        if (l.type === 'desc') sortedData.reverse();
+      }
+    });
+
+    return sortedData;
+  };
+
+  const paginateData = (data, page, pageSize) => {
+    const start = (page - 1) * pageSize;
+    const end = start + pageSize;
+    const paginatedData = data.slice(start, end);
+    return paginatedData;
+  };
+
+  const data = ${JSON.stringify(data.slice(0, 10), null, 4)};
+  const [formattedData, setFormattedData] = React.useState(data);
 
   const schema = [
     {
-      name: 'firstName',
+      name: 'name',
       displayName: 'Name',
-      cellType: 'AVATAR_WITH_TEXT',
-      width: '25%',
-      translate: (a) => ({
+      width: 300,
+      resizable: true,
+      separator: true,
+      tooltip: true,
+      translate: a => ({
         title: \`\${a.firstName} \${a.lastName}\`,
         firstName: a.firstName,
-        lastName: a.lastName,
+        lastName: a.lastName
       }),
+      filters: [
+        { label: 'A-G', value: 'a-g' },
+        { label: 'H-R', value: 'h-r' },
+        { label: 'S-Z', value: 's-z' },
+      ],
+      onFilterChange: (a, filters) => {
+        for (const filter of filters) {
+          switch (filter) {
+            case 'a-g':
+              if (a.firstName[0].toLowerCase() >= 'a' && a.firstName[0].toLowerCase() <= 'g') return true;
+              break;
+            case 'h-r':
+              if (a.firstName[0].toLowerCase() >= 'h' && a.firstName[0].toLowerCase() <= 'r') return true;
+              break;
+            case 's-z':
+              if (a.firstName[0].toLowerCase() >= 's' && a.firstName[0].toLowerCase() <= 'z') return true;
+              break;
+          }
+        }
+        return false;
+      },
+      cellType: 'AVATAR_WITH_TEXT',
     },
     {
       name: 'email',
       displayName: 'Email',
-      width: '25%',
+      width: 350,
+      resizable: true,
+      sorting: false,
+      cellType: 'WITH_META_LIST'
     },
     {
       name: 'gender',
       displayName: 'Gender',
-      width: '25%',
+      width: 200,
+      resizable: true,
+      comparator: (a, b) => a.gender.localeCompare(b.gender),
+      cellType: 'STATUS_HINT',
+      translate: a => ({
+        title: a.gender,
+        statusAppearance: (a.gender === 'Female') ? 'alert' : 'success'
+      }),
+      filters: [
+        { label: 'Male', value: 'male' },
+        { label: 'Female', value: 'female' },
+      ],
+      onFilterChange: (a, filters) => {
+        for (const filter of filters) {
+          if (a.gender.toLowerCase() === filter) return true;
+        }
+        return false;
+      },
     },
     {
-      name: 'status',
-      displayName: 'Status',
-      width: '25%',
+      name: 'icon',
+      displayName: 'Icon',
+      width: 100,
+      resizable: true,
+      align: 'center',
+      cellType: 'ICON',
+      sorting: false,
+      translate: _ => ({
+        icon: 'event'
+      })
+    },
+    {
+      name: 'customCell',
+      displayName: 'Custom Cell',
+      width: 200,
+      resizable: true,
+      cellType: 'WITH_META_LIST',
+      sorting: false,
+      cellRenderer: props => {
+        return (
+          <>
+            <Icon className="mr-5" name="event" />
+            <GridCell
+              {...props}
+              schema={{
+                ...props.schema,
+                name: 'email'
+              }}
+            />
+          </>
+        );
+      }
     },
   ];
 
+  const fetchData = (options) => {
+    const {
+      page,
+      pageSize,
+      sortingList,
+      filterList,
+      searchTerm
+    } = options;
+
+    const onSearch = (d, searchTerm = '') => {
+      return (
+        d.firstName.toLowerCase().match(searchTerm.toLowerCase())
+        || d.lastName.toLowerCase().match(searchTerm.toLowerCase())
+      );
+
+      return true;
+    }
+
+    const filteredData = filterData(schema, data, filterList);
+    const searchedData = filteredData.filter(d => onSearch(d, searchTerm));
+    const sortedData = sortData(schema, searchedData, sortingList);
+    setFormattedData(sortedData);
+
+    if (page && pageSize) {
+      return new Promise(resolve => {
+        window.setTimeout(() => {
+          const start = (page - 1) * pageSize;
+          const end = start + pageSize;
+          const slicedData = sortedData.slice(start, end);
+          resolve({
+            searchTerm,
+            schema,
+            count: sortedData.length,
+            data: slicedData,
+          });
+        }, 2000);
+      });
+    }
+
+    return new Promise(resolve => {
+      window.setTimeout(() => {
+        resolve({
+          searchTerm,
+          schema,
+          count: sortedData.length,
+          data: sortedData,
+        });
+      }, 2000);
+    });
+  }
+
+  const loaderSchema = ${JSON.stringify(loaderSchema, null, 4)};
+
+  const onDataExport = (data) => {
+    console.log("Exporting data", data);
+  }
+
+  const globalActionTrigger = (data) => {
+    return (<Button onClick={() => onDataExport(data)}>Export</Button>);
+  } 
+
+  const selectionActionRenderer = (selectedData, selectAll) => {
+    console.log('selectedData in output', selectedData, 'selectAll', selectAll);
+    return (
+      <div className="d-flex align-items-center">
+        <Button size="tiny" className="mr-4">Delete</Button>
+        <Button size="tiny">Export</Button>
+      </div>
+    )
+  }
+
   return (
-    <div className="vh-50">
+    <div className="vh-75">
       <Card className="h-100 overflow-hidden">
         <Table
-          page={1}
-          data={data}
-          schema={schema}
-          showMenu={true}
+          loaderSchema={loaderSchema}
+          fetchData={fetchData}
           withHeader={true}
-          withPagination={true}
-          pageSize={4}
+          uniqueColumnName="firstName"
           headerOptions={{
+            selectionActionRenderer,
             withSearch: true,
+            globalActionRenderer : globalActionTrigger,
+            allowSelectAll: true,
           }}
+          withCheckbox={true}
+          onSelect={(rowIndex, selected, selectedList, selectAll) => console.log(\`on-select: - rowIndex: \${ rowIndex } selected: \${ selected } selectedList: \${ JSON.stringify(selectedList) } selectAll: \${ selectAll } \`)}
+          withPagination={true}
+          pageSize={5}
+          onPageChange={newPage => console.log(\`on-page-change:- \${newPage}\`)}
         />
       </Card>
     </div>
   );
-}`;
+};
+`;
 
 export default {
   title: 'Components/Table/All',
@@ -261,6 +324,7 @@ export default {
     docs: {
       docPage: {
         customCode,
+        title: 'All',
         props: {
           components: { AsyncTable, SyncTable },
         },
